% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_phase_estimators.R
\name{est_tot_ht}
\alias{est_tot_ht}
\title{Estimate the population total using the Horvitz-Thompson estimator}
\usage{
est_tot_ht(y, prob, Nd = NULL)
}
\arguments{
\item{y}{Study variable}

\item{prob}{Inlcusion probabilities of the design}

\item{Nd}{Domain size. If provided a more precise estimator for the domain
total is used.}
}
\value{
A named vector of three elements containing the estimated total, the
 sample size, and the number of non-zero sample units
}
\description{
This function estimates the population total of a study variable using the
Horvitz-Thompson estimator. The function also allows for domain estimation.
Two cases are distinguished: domian size is known and domain size is unknown.
}
\references{
See Särndal et al. (1992), eq. 10.3.3 and 10.3.6
}

