% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase_estimators.R
\name{est_ts_ma_srswor}
\alias{est_ts_ma_srswor}
\title{Model-assisted estimators for two-stage simple random sampling}
\usage{
est_ts_ma_srswor(dt_samp, NI, nI, id_psu = "id", y = quote(y),
  N_i = quote(N_i), n_i = quote(n_i), res = quote(res),
  t_pred = quote(t_pred))
}
\arguments{
\item{NI}{Number of PSUs}

\item{nI}{PSU sample size}

\item{y}{Quoted name of the variable storing the observed values of the study
variable.}

\item{N_i}{quoted name of the variable storing the size of the PSUs}

\item{n_i}{quoted name of the variable storing the number of selected SSUs in
each PSUs}

\item{res}{Quoted name of the variable stroing the residuals between observed
and predicted values.}

\item{t_pred}{Quoted name of the variable storing predicted PSU totals}

\item{dt_sample}{A data.table object containing information about the PSUs and
SSUs selected in the two sampling stages.}

\item{key_var}{Name of the variable identifying the PSUs. Must be present in
both data.tables.}
}
\value{
A named list containing estimates for the total and its variance
 estimate
}
\description{
The population is separated into a set of contiguous PSUs (clusters/strips),
which are sampled in the first stage. SSUs are selected from the selected PSUs
in the second stage independently. The estimators correspond to case C in
Särndal et al. (1992), where auxiliaries are avalaible for SSUs in the
selected PSUs. Modelling is done at the element level.
}
\details{
See Särndal et al. (1992), equation 8.9.7 for the estimation of the
 population total and Equation 8.9.27 for traditional variance estimation.
 Under some circumstances, the variance estimator may deliver negative
 estimates. Therefore, two alternative variance estimators are returned as
 well. The first is the so called prediction-based, model-assisted variance
 estimator and uses model-assisted estimates of PSU totals instead of
 Horvitz-Thompson PSU total estimates. This variance estimator should be more
 stable (Saarela et al. 2016). The second ignores the variation that comes
 from estimating PSU totals (Särndal et al., p. 154, Equation 4.6.2).
}
\references{
Särndal, C. E., B. Swensson and J. Wretman (1992). Model assisted
 survey sampling. New York, Springer-Verlag.

 Saarela, S.; Andersen, H.-E.; Grafstrom, A.; Schnell, S.; Gobakken, T.;
 Næsset, E.; Nelson, R. F.; McRoberts, R. E.; Gregoire, T. & Ståhl, G. A new
 prediction-based variance estimator for two-stage model-assisted surveys of
 forest resources. Remote Sensing of Environment, 2016
}

