% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase_estimators.R
\name{est_ts_mb}
\alias{est_ts_mb}
\title{Model-based inference for two-stage surveys}
\usage{
est_ts_mb(dt_samp, X, parm_cov, t_pred = quote(t_pred), N, NI, nI,
  N_i = quote(N_i), a_psu)
}
\arguments{
\item{X}{Matrix containing the auxiliary variables (observations stored in
rows, variables in columns). Can be a column vector of auxiliary averages in
case linear models are used.}

\item{parm_cov}{Model parameter covariance matrix.}

\item{t_pred}{Quoted name of the variable storing predicted PSU totals}

\item{N}{Population size (Numbr of SSUs in the population).}

\item{NI}{Number of PSUs}

\item{nI}{PSU sample size}

\item{dt_sample}{A data.table object containing information about the PSUs and
SSUs selected in the two sampling stages.}
}
\value{
A named list containing estimates for the total and its variance
 estimate
}
\description{
The population is separated into a set of contiguous PSUs (clusters/strips),
which are sampled in the first stage. SSUs are selected from the selected PSUs
in the second stage independently.
}
\details{
See Ståhl et al. (2011).
}
\references{
Ståhl, G., S. Holm, T. G. Gregoire, T. Gobakken, E. Næsset and R.
 Nelson (2011). "Model-based inference for biomass estimation in a LiDAR
 sample survey in Hedmark County, Norway." Canadian Journal of Forest
 Research 41: 96-107.
}

