% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase_estimators.R
\name{est_ts_srswor}
\alias{est_ts_srswor}
\title{Horvitz-Thompson estimators for two-stage simple random sampling}
\usage{
est_ts_srswor(dt_samp, NI, nI, id_psu = "id", y = quote(y),
  N_i = quote(N_i), n_i = quote(n_i))
}
\arguments{
\item{NI}{Number of PSUs}

\item{nI}{PSU sample size

 @details See S?rndal et al. (1992), equation 4.3.21 for the estimation of
 the population total and Equation 4.3.22 for variance estimation. The
 variance estimator may under some circumstances deliver negative estimates.
 Therefore, an alternative variance estimator that ignores the variation that
 comes from estimating PSU totals is provided as an alternative (Särndal et
 al., p. 154, Equation 4.6.2).}

\item{y}{Quoted name of the variable storing the observed values of the study
variable.}

\item{N_i}{quoted name of the variable storing the size of the PSUs}

\item{n_i}{quoted name of the variable storing the number of selected SSUs in
each PSUs}

\item{dt_sample}{A data.table object containing information about the PSUs and
SSUs selected in the two sampling stages.}

\item{key_var}{Name of the variable identifying the PSUs. Must be present in
both data.tables.}
}
\value{
A named list containing estimates for the total and its variance
 estimate
}
\description{
The population is separated into a set of contiguous PSUs (clusters/strips),
which are sampled in the first stage. SSUs are selected from the selected PSUs
in the second stage independently.
}
\references{
Särndal, C. E., B. Swensson and J. Wretman (1992). Model assisted
 survey sampling. New York, Springer-Verlag.
}

