% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_var_mb}
\alias{est_var_mb}
\title{Model-based variance estimator for an estimated population total}
\usage{
est_var_mb(cov, X)
}
\arguments{
\item{cov}{Model parameter covariance matrix}

\item{X}{Design matrix}
}
\value{
The variance estimate
}
\description{
Model-based variance estimator for an estimated population total
}

