% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_phase_estimators.R
\name{est_var_srswor}
\alias{est_var_srswor}
\title{Estimate the variance of the estimated population total under simple random
sampling without replacement}
\usage{
est_var_srswor(y, n, N, Nd = NULL)
}
\arguments{
\item{y}{Study variable}

\item{n}{Sample size. May be larger than the length of y if the aim is to
estimate the variance for a certain domain.}

\item{N}{Size of the population}

\item{Nd}{Size of the domain. If provided a more precise estimator for the
variance of the total is used.}
}
\value{
A named vector of one element containing the variance estimate.
}
\description{
This function estimates the variance of the estimated population total of a
study variable under simple random sampling without replacement. The function
also allows for domain estimation if domain size is either known or unknown.
}
\references{
Särndal et al. (1992), eq. 10.3.10 and 10.3.12
}

