% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_phase_estimators.R
\name{est_var_srswor_ma}
\alias{est_var_srswor_ma}
\title{Estimate the variance of the estimated population total under simple random
sampling without replacement followign generalised regression estimation}
\usage{
est_var_srswor_ma(res, N)
}
\arguments{
\item{res}{Residuals between observed and predicted values in the sample}

\item{N}{Size of the population. For domain variance provide the domain size.}
}
\value{
A named vector of one element containing the variance estimate.
}
\description{
This function estimates the variance of the estimated population total of a
study variable under simple random sampling without replacement when
generalised linear regression was used. The function considers domain
estimation.
}
\references{
Särndal and Hidiroglou (1989)
}

