% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase_estimators.R
\name{est_var_succ_diff}
\alias{est_var_succ_diff}
\title{Succesive difference estimator to estimate variance of a systematic sample}
\usage{
est_var_succ_diff(y, N, n, w = NULL)
}
\arguments{
\item{y}{Observations}

\item{N}{Population size}

\item{n}{Sample size}

\item{w}{Weights in case systematic sampling of clusters of unequal size is
applied. See details for how they are calculated.}
}
\value{
An atomic vector with the variance estimate.
}
\description{
Succesive difference estimator to estimate variance of a systematic sample
}
\details{
For the unweighted version, see the \eqn{v_4}-estimator in Wolter
  (2003, chapter 8.3). For the weighted version and the case of cluster of
  unequal size, see Nelson et al. (2008, Equation 3). The cluster weights are
  defined as \deqn{W_i = N_i(\sum_{i=1}^{n-1}N_i),} where \eqn{N_i} denotes
  the size of a cluster and \eqn{n} is the number of clusters selected into
  the sample.
}
\references{
Nelson, R., Easset, T. Gobakken, G. Stahl and T. G. Gregoire
(2008). "Regional Forest Inventory using an Airborne Profiling LiDAR."
Journal of Forest Planning 13: 287-294.

Wolter, K. (2003). Introduction to Variance Estimation, Springer.
}

