% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase_estimators.R
\name{get_psu_stats}
\alias{get_psu_stats}
\title{Internal helper function to prepare data from a two-stage sample for
estimation by other functions}
\usage{
get_psu_stats(dt_samp, id_psu, y, res = NULL)
}
\arguments{
\item{dt_samp}{A data.table object containing information about the PSUs and
SSUs selected in the two sampling stages.}

\item{y}{Quoted name of the variable storing the observed values of the study
variable}

\item{res}{Quoted name of the variable under which residuals between observed
and predicted values are stored. Optional in case of model-assisted
estimation}

\item{key_var}{Name of the variable identifying the PSUs.}
}
\value{
A data.table object containing PSU summaries needed for estimation by
  other functions.
}
\description{
Internal helper function to prepare data from a two-stage sample for
estimation by other functions
}

