% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{res_ss_hc_tm}
\alias{res_ss_hc_tm}
\title{Return the squared estimated residuals when two OLS models are
applied in model-based inference and heteroscedasticity is present}
\usage{
res_ss_hc_tm(X, Z, beta, alpha, vcov_beta_hc)
}
\arguments{
\item{X}{Design matrix of the first model for the first-phase sample}

\item{Z}{Design matrix of the second model for the first-phase sample}

\item{beta}{Parameter estimates of the first model (column vector)}

\item{alpha}{Parameter estimates of the second model (column vector)}

\item{vcov_beta_hc}{Variance/co-variance matrix of first-model parameters
estimates}
}
\value{
The residual sum of squares
}
\description{
Return the squared estimated residuals when two OLS models are
applied in model-based inference and heteroscedasticity is present
}

