% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_fix_sum.R
\name{round_fix_sum}
\alias{round_fix_sum}
\title{Round to integer keeping the sum fixed}
\usage{
round_fix_sum(x)
}
\arguments{
\item{x}{A numeric vector which must sum to an integer}
}
\value{
A vector of rounded integers summing up to sum(x).
}
\description{
Round to integer keeping the sum fixed
}
\references{
Code taken from roundfixS function of sfsmisc-package by Martin
  Maechler
}

