% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcov_tm_hc}
\alias{vcov_tm_hc}
\title{Estimated model parameter covariance matrix for the top-level model in
hierarchical model-based inference setting when two models are applied}
\usage{
vcov_tm_hc(res_sq, X)
}
\arguments{
\item{Z}{Design matrix of the second model for the first-phase sample}

\item{res}{Sum of squared residuals for the top-level model}
}
\value{
Estimated covariance matrix for model parameter estimates
}
\description{
Estimated model parameter covariance matrix for the top-level model in
hierarchical model-based inference setting when two models are applied
}
\details{
The heteroskedasticity-consistent version
}

