% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_utils.R
\name{vcov_white}
\alias{vcov_white}
\title{White's heteroskedasticity consistent covariance matrix estimator}
\usage{
vcov_white(X, res)
}
\arguments{
\item{X}{Design matrix}

\item{res}{Model residuals}
}
\value{
Estimated parameter variance-co-variance matrix
}
\description{
White's heteroskedasticity consistent covariance matrix estimator
}
\references{
White, H. (1980). "A Heteroskedasticity-Consistent Covariance
  Matrix Estimator and a Direct Test for Heteroskedasticity." Econometrica
  48(4): 817-838.
}

